fix: adjust horizontal paging to changes in UiMenu
