+++ src/components/organisms/UiHorizontalPaging/_internal/UiHorizontalPagingItem.vue
@@ -10,13 +10,11 @@ import {
   onBeforeUnmount,
   type Ref,
   type ComputedRef,
-  useAttrs,
 } from 'vue';
-import UiButton from '../../../atoms/UiButton/UiButton.vue';
 import type { HorizontalPagingHandleItems } from '../UiHorizontalPaging.vue';
-import type { HTMLTag } from '../../../../types';
+import type { MenuItemProps } from '../../UiMenu/_internal/MenuItemProps';
 
-export interface HorizontalPagingItemProps {
+export interface HorizontalPagingItemProps extends MenuItemProps {
   /**
    * Use this props to set inside pages item label.
    */
@@ -29,28 +27,16 @@ export interface HorizontalPagingItemProps {
    * Use this props to set inside pages item name.
    */
   name?: string;
-  /**
-   * Use this props to set list item content tag.
-   */
-  tag?: HTMLTag;
 }
 defineOptions({ inheritAttrs: false });
 const props = withDefaults(defineProps<HorizontalPagingItemProps>(), {
   label: '',
   title: '',
   name: '',
-  tag: UiButton,
 });
-const attrs = useAttrs();
 const activeItemName = inject<ComputedRef<string>>('activeItemName', computed(() => ''));
 const isActive = computed(() => activeItemName.value === props.name);
-const item = computed(() => ({
-  label: props.label,
-  title: props.title,
-  name: props.name,
-  ...(props.tag === UiButton ? {} : { tag: props.tag }),
-  ...attrs,
-}));
+const item = computed(() => ({ ...props }));
 const items = inject<Ref<HorizontalPagingHandleItems>>('items', ref({}));
 items.value[props.name] = item.value;
 onBeforeUnmount(() => {
