+++ src/components/organisms/UiMenu/_internal/MenuItemProps.ts
@@ -1,0 +1,3 @@
+import type { ListItemProps } from '../../UiList/_internal/ListItemProps';
+
+export interface MenuItemProps extends ListItemProps {}
